forwardport: make socket handling more reliable
make the socket non-blocking, so things don't go awry when a client
freezes. the socket reading will be able to deal with that. the
socket writing function will simply ignore errors - that's not a
problem, as the socket's internal buffer is big enough for our
payloads, so blocking would indicate a serious problem anyway.

set the accept() queue length to 10. no reason to limit it to one.

speculative fix for bug 208979

svn path=/trunk/KDE/kdebase/runtime/; revision=1196642