Do not require to have a DISPLAY env variable if WAYLAND_DISPLAY is set
If kdesu is compiled with X11 it required the DISPLAY variable to be
set. This is no longer correct as it might have been compiled with
X11 but is run on Wayland. Thus the code checks now also for
WAYLAND_DISPLAY in the HAVE_X11 ifdef blocks. The Wayland support
should become more complete, I do not know how it behaves if we compile
without X11 support. Unfortunately there are no autotests and no test
applications which one could use.

REVIEW: 115717