SuProcess: Disable echo in the PTY before starting sudo
Recent versions of sudo have use_pty enabled by default, which means that sudo
creates a PTY for starting the user process inside after successful
authentication. This PTY inherits the configuration of sudo's TTY, but later
changes are not transferred. Make sure that echo is already disabled when sudo
is started, as disabling it later has no effect on the nested PTY.

BUG: 452532